chore - fake object 제거
