+++ src/test/java/kitchenpos/bo/MenuBoTestWithBuilder.java
@@ -1,0 +1,4 @@
+package kitchenpos.bo;
+
+public class MenuBoTestWithBuilder {
+}
