+++ src/test/java/kitchenpos/bo/ProductBoTest.java
@@ -1,14 +1,12 @@
 package kitchenpos.bo;
 
-import kitchenpos.dao.DefaultProductDao;
+import kitchenpos.dao.ProductDao;
 import kitchenpos.model.Product;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
@@ -19,7 +17,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.BDDMockito.given;
 
 class ProductBoTest extends MockTest{
-    @Mock private DefaultProductDao productDao;
+    @Mock private ProductDao productDao;
     @InjectMocks private ProductBo productBo;
 
     private Product product;
