+++ src/test/java/kitchenpos/builder/MenuBuilder.java
@@ -1,0 +1,4 @@
+package kitchenpos.builder;
+
+public class MenuBuilder {
+}
