+++ src/test/java/kitchenpos/builder/MenuProductBuilder.java
@@ -1,0 +1,4 @@
+package kitchenpos.builder;
+
+public class MenuProductBuilder {
+}
