+++ src/test/java/kitchenpos/builder/ProductBuilder.java
@@ -1,0 +1,4 @@
+package kitchenpos.builder;
+
+public class ProductBuilder {
+}
